Exctract interfaces from repository
Seperate implementation and interface of repository.
