+++ eatgo-api/src/main/java/kr/co/darcie/eatgo/domain/RestaurantRepository.java
@@ -1,33 +1,9 @@
 package kr.co.darcie.eatgo.domain;
 
-import org.springframework.stereotype.Component;
-
-import java.util.ArrayList;
 import java.util.List;
 
-@Component
-public class RestaurantRepository {
+public interface RestaurantRepository {
+    List<Restaurant> findAll();
 
-    private List<Restaurant> restaurants = new ArrayList<>();
-
-    public RestaurantRepository() {
-
-        restaurants.add(new Restaurant(1004L, "Bob zip", "Seoul"));
-        restaurants.add(new Restaurant(2020L, "Cyber Food", "Seoul"));
-
-    }
-
-
-    public List<Restaurant> findAll() {
-
-        return restaurants;
-    }
-
-    public Restaurant findById(Long id) {
-
-        return restaurants.stream()
-                .filter(r->r.getId().equals(id))
-                .findFirst()
-                .orElse(null);
-    }
+    Restaurant findById(Long id);
 }
