+++ eatgo-api/src/main/java/kr/co/darcie/eatgo/interfaces/RestaurantController.java
@@ -1,21 +1,19 @@
 package kr.co.darcie.eatgo.interfaces;
 
 import kr.co.darcie.eatgo.domain.Restaurant;
-import kr.co.darcie.eatgo.domain.RestaurantRepository;
+import kr.co.darcie.eatgo.domain.RestaurantRepositoryImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 
 @RestController
 public class RestaurantController {
 
     @Autowired
-    private RestaurantRepository repository;
+    private RestaurantRepositoryImpl repository;
 
     @GetMapping("/restaurants")
     public List<Restaurant> list() {
